Code from Elm Orlando's November 2017 meeting
Posting the code as-is from what we worked with in our mob programming
session. Could use some refactoring for parts of this.

The current status is that the ball is moving, the player paddle is
adjustable with the W and S keys, the "AI" paddle is adjustable with the
O and K keys. The scores are also incrementing when the ball strikes the
edges, but there's currently no collision detection for the paddles.